envs/slideforge_env/server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
Pillow>=10.0.0
pydantic>=2.0
